changed long long to int64_t

+++ maths/ceil_div.cpp
@@ -5,6 +5,6 @@ using namespace std;
 /* Ceil division */
 
 // TC - O(1), SC - O(1)
-long long ceil_div(long long a, long long b) {
+int64_t ceil_div(int64_t a, int64_t b) {
   return (a + b - 1) / b;
 }
